requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pyjwt==2.8.0
qrcode==7.4.2
pillow==10.0.1
python-dotenv==1.0.1
